Rework of the cphase interpolation
The closure phase interpolation was not correctly implemented, now for all uv combinations the closure phases are summed up after being interpolated.

Still the question if the int roll is better (Jacob's method) or the interpolation (Anthony's method) and if interpolation is better with "linear" or there is a better method?